+++ src/images.h
@@ -34,7 +34,7 @@ struct CompressOptions {
 };
 
 void optimizeImages(QPDF &qpdf, const CompressOptions &opts);
-void downscaleImages(QPDF &qpdf, int maxDpi);
+void downscaleImages(QPDF &qpdf, int maxDpi, int quality);
 void deduplicateImages(QPDF &qpdf);
 void optimizeExistingJpegs(QPDF &qpdf);
 void stripMetadata(QPDF &qpdf);